+++ src/cuda/gpu_ray_class.cuh
@@ -3,40 +3,40 @@
 
 class Ray
 {
-    double wepl;
-    double energy;
-    double initial_energy;
+    float wepl;
+    float energy;
+    float initial_energy;
     short  beam_id;
     short  spot_id;
 public:
-    double3 pos;
-    double3 dir;
+    float3 pos;
+    float3 dir;
 
     // CONSTRUCTORS ----------------------------------------------------------
-    __device__ Ray (double4 x_, double4 vx_, short2 ix);
+    __device__ Ray (float4 x_, float4 vx_, short2 ix);
 
     // BASIC GETTERS ---------------------------------------------------------
-    __device__ double3 get_position () const;
-    __device__ double3 get_direction () const;
+    __device__ float3 get_position () const;
+    __device__ float3 get_direction () const;
     __device__ short get_beam_id () const;
     __device__ short get_spot_id () const;
-    __device__ double get_energy () const;
-    __device__ double get_initial_energy () const;
-    __device__ double get_wepl () const;
+    __device__ float get_energy () const;
+    __device__ float get_initial_energy () const;
+    __device__ float get_wepl () const;
     // GETTERS ---------------------------------------------------------------
     __device__ bool is_alive ();
     // BASIC SETTERS ---------------------------------------------------------
-    __device__ void set_energy (double e);
-    __device__ void set_wepl (double e);
-    __device__ void set_position (double4 d);
-    __device__ void set_position (double3 d);
-    __device__ void set_direction (double4 d);
-    __device__ void set_direction (double3 d);
+    __device__ void set_energy (float e);
+    __device__ void set_wepl (float e);
+    __device__ void set_position (float4 d);
+    __device__ void set_position (float3 d);
+    __device__ void set_direction (float4 d);
+    __device__ void set_direction (float3 d);
     // SETTERS ---------------------------------------------------------------
-    __device__ void set_direction_to_point (double3 d);
-    __device__ void set_direction_to_point (double4 d);
+    __device__ void set_direction_to_point (float3 d);
+    __device__ void set_direction_to_point (float4 d);
     // ACTIONS ---------------------------------------------------------------    
-    __device__ void move (const double& step, const double& wepl, const double& de);
+    __device__ void move (const float& step, const float& wepl, const float& de);
     __device__ void kill();
     // QUERIES ---------------------------------------------------------------
     __device__ void print();
@@ -44,7 +44,7 @@ public:
 
 private:
     bool  _alive     = true;
-    double _min_wepl = 1e-08;
+    float _min_wepl = 1e-08;
 };
