+++ src/gpu_main.cuh
@@ -17,32 +17,32 @@ void initialize_device (cudaEvent_t& start);
 void stop_device (cudaEvent_t& start);
 
 void gpu_raytrace (const Patient_Parameters_t& pat,
-                   Array4<double>& endpoints,
+                   Array4<float>& endpoints,
                    std::string output_file,
-                   const Array4<double>& orig_endpoints = Array4<double>());
+                   const Array4<float>& orig_endpoints = Array4<float>());
 
 void gpu_raytrace_original (const Patient_Parameters_t &pat,
                             const Volume_t &ct,
-                            Array4<double>& endpoints,
-                            Array4<double>& init_pos,
-                            Array4<double>& init_pat_pos,
+                            Array4<float>& endpoints,
+                            Array4<float>& init_pos,
+                            Array4<float>& init_pat_pos,
                             const Parser& parser,
-                            Array4<double>& influence);
+                            Array4<float>& influence);
 
 void gpu_raytrace_warped (const Patient_Parameters_t &pat,
                           const Volume_t &ct,
-                          const Array4<double>& orig_endpoints,
-                          const Array4<double>& init_pos,
-                          Array4<double>& endpoints,
+                          const Array4<float>& orig_endpoints,
+                          const Array4<float>& init_pos,
+                          Array4<float>& endpoints,
                           const Parser& parser,
-                          Array4<double>& influence);
+                          Array4<float>& influence);
 
 void gpu_calculate_influence (const uint& nspots,
                               const uint& nprobes,
-                              Array4<double>& influence,
+                              Array4<float>& influence,
                               std::vector<float>& spot_weights,
                               std::vector<float>& inf_volume,
-                              const std::vector<double>& new_energies = std::vector<double>());
+                              const std::vector<float>& new_energies = std::vector<float>());
 
 void printDevProp (const int device, bool verbose);
